- Clean up code and create sagas.js for sync bg comments
